// frontend/src/app/models/user.ts
export default class User {
    username: string = "";
    password: string = "";
    name: string = "";
    surname: string = "";
    sex: string = "";
    address: string = "";
    phone_number: string = "";
    email: string = "";
    profile_photo: string = "";
    credit_card_number: string = "";
    role: string = "";
    approved: boolean = false;
    rejected: boolean = false;
}